fix a code reload  concurency issue
I started getting the following error:
```
[GRPC] app err:#<GRPC::ActiveCall:0x000000011501e260>, status:13:please, finish your configuration and call Zeitwerk::Loader#setup once all is ready
```

When Zeitwerk does a `reload`, it "un-setups" (`unload`) itself before a
rebuild of itself (`setup`). But it's not thread-safe so if two threads
get in there, if the second one hits reload before the first has
finished to re-setup. Then it's going to detect that setup hasn't run
and throw the ` finish your configuration and call Zeitwerk::Loader#setup`
error.

I think i started getting this after the following gem udpate(?):

```
google-protobuf (3.21.9)
grpc (1.50.0)
grpc-tools (1.50.0)
```

Maybe there were changes in concurency in those gems.
Maybe it's something else, got lots of changes merging on this repo. But
it fixed my app so I'm using this for now.